Create optimized_map.py
this app does the same function as "generatemap.py" but it optimizes the map display using the FastMarkerCluster library in order to render the map more quickly. This saves the user time when viewing the map online. Some features are lost when using this optimized version of the app - namely the map legend differentiating each coordinate. Users can still see individual coordinate data by clicking on the coordinate bubbles.